CHORE: title & input 컴포넌트 분리
