+++ 0617/src/components/Todolist/titleAndAddTask.module.scss
@@ -1,0 +1,10 @@
+.addTaskList {
+  width: 100%;
+  display: flex;
+  background-color: pink;
+  margin-top: 20px;
+}
+.addTaskInput {
+  flex: 1;
+  padding: 10px;
+}
