+++ 0617/src/page/todolist/todolist.module.scss
@@ -11,16 +11,6 @@
   border-left: 1px solid #ddd;
   padding: 20px;
 }
-.addTaskList {
-  width: 100%;
-  display: flex;
-  background-color: pink;
-  margin-top: 20px;
-}
-.addTaskInput {
-  flex: 1;
-  padding: 10px;
-}
 
 .taskListUi {
   margin-top: 20px;
